fix(be): 리더보드 쿼리 버그 수정 — bigint 캐스트, role 필터 제거, getMySoloRank 일관성
- ROW_NUMBER() - 1 반환값이 bigint → date - bigint 연산 오류: ::int 캐스팅 추가
- getMySoloRank 전체/언어 브랜치 role = 'USER' 필터 추가 → 리더보드와 일관성 확보
- getSoloLeaderboard/Streak role 필터 제거 → ADMIN 포함 전체 유저 랭킹

+++ src/snippet-result/snippet-result.repository.ts
@@ -355,7 +355,7 @@ export class SnippetResultRepository {
             ),
             numbered AS (
                 SELECT day,
-                       ROW_NUMBER() OVER (ORDER BY day DESC) - 1 AS rn
+                       (ROW_NUMBER() OVER (ORDER BY day DESC) - 1)::int AS rn
                 FROM daily
                 WHERE day <= (SELECT base_day FROM base)
             ),
